requirements.txt
autopep8==2.0.1
blinker==1.5
boto3==1.26.61
botocore==1.29.61
click==8.1.3
flake8==6.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==2.0.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
